🎨 extends hook to generic useElementDimention
