+++ uikit/Table/index.tsx
@@ -12,7 +12,7 @@ import { StyledTable, StyledTableProps } from './styledComponent';
 import TablePagination from './TablePagination';
 import DefaultNoDataComponent from './NoDataComponent';
 import { TableProps } from 'react-table';
-import useParentWidth from './useParentWidth';
+import useElementWidth from '../utils/Hook/useElementDimention';
 
 export { default as TablePagination, TableActionBar } from './TablePagination';
 
@@ -87,6 +87,7 @@ function Table<Data = { [k: string]: any }>({
     }
   },
   parentRef,
+  withResizeBlur = false,
   ...rest
 }: Partial<TableProps<Data>> & {
   variant?: TableVariant;
@@ -96,6 +97,7 @@ function Table<Data = { [k: string]: any }>({
   primaryKey?: string;
   columns: TableProps<Data>['columns']; //columns is required
   parentRef: React.RefObject<HTMLElement>;
+  withResizeBlur?: boolean;
 } & StyledTableProps) {
   const TrComponent = rest.TrComponent || DefaultTrComponent;
   const getTrProps = rest.getTrProps || (() => ({}));
@@ -107,14 +109,14 @@ function Table<Data = { [k: string]: any }>({
 
   // react-table needs an explicit pixel width to handle horizontal scroll properly.
   // This syncs up the component's width to its container.
-  const { width, resizing } = useParentWidth(parentRef);
+  const { width, resizing } = useElementWidth(parentRef);
 
   return (
     <StyledTable
       style={{
         // this is written with style object because css prop someone only applies to the header
         transition: 'all 0.25s',
-        filter: resizing ? 'blur(8px)' : 'blur(0px)',
+        filter: resizing && withResizeBlur ? 'blur(8px)' : 'blur(0px)',
         width,
       }}
       withRowBorder={withRowBorder}
